running_tasks now counts across all nodes.
running_tasks used client.containers.get(id) to verify that container
had good status which works only on the local host. Classic mistake.
It can also be noted that containers can say "running" while the task
is "starting". Since the main use case for running_tasks is to be part
of the steady-state hypothesis, it can be argued that "starting" !=
"running".